Add game tick code
Game tick is in two parts: Draw and update (because of old habits where
these would fire separately)

This is all broken up a lot, so it's not all that performant - But I doubt
it will cause problems for something like this. If it does, I will start
optimizing then.