Give a one-GameObject gap between statics and blasts
This fixes #5.... Sortof. It throws a tarp over it. I don't really know
why #5 was happening, so this will do for now.

This is sort of like the tarp thrown over the "second rotated object
causes infinite loop" bug, where the rotated sprite code in RotateBlit was
corrupting esp/ebp, and I just saved/restored them from known parts of
memory whenever the function ran. Which is, by the way, still an issue.