Upgrade paused counter to a doubleword
The reason why the game wasn't staying paused was that `paused`, which was
originally a flag declared as a byte to save space, was overflowing.

To prevent this, we simply allocate more space to it.

A word would allow a pause of about 1.2 hours, which is good, and probably enough, but not definitely enough.

A doubleword allows a pause of about 4.5 years. And frankly, I'd be amazed
if anyone is playing this game in 4.5 years, nevermind pausing it for that
long.