Rapidfire balance changes: More shots, less speed
The fire rate is increased from ~2 shots per second to 3 shots per second.
In exchange, rapidfire shots move half as fast as they used to.

This change permits the blaster to still require some skill to hit
targets. Furthermore, at the old speed, the sprite was moving much too
fast to be perceived as motion very well.

If this is still wrong once the powerup is no longer indefinite in
length, I'll reduce the velocity to be around half that of the normal
blaster, with the realistic-ish explanation of blaster overload reducing
output power or something.